+++ gatsby-config.ts
