+++ package.json
@@ -56,12 +56,15 @@
     "webpack": "^5.51.2"
   },
   "devDependencies": {
+    "@types/node": "^17.0.36",
+    "@types/react-dom": "^18.0.5",
     "eslint": "^8.16.0",
     "eslint-plugin-react": "^7.25.1",
     "husky": "^8.0.1",
     "lint-staged": "^12.4.2",
     "prettier": "^2.3.0",
-    "pretty-quick": "^3.1.1"
+    "pretty-quick": "^3.1.1",
+    "typescript": "^4.7.2"
   },
   "repository": {
     "type": "git",
