+++ src/components/Button.tsx
@@ -3,7 +3,7 @@ import { Link } from "gatsby";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import { color } from "@theme/config";
+import { color } from "../theme/config";
 
 const StyledButton = styled(Link)`
   background-image: linear-gradient(
