+++ src/components/EventsDeck.tsx
