+++ src/templates/blogTemplate.tsx
